Move specialization-related test to FULL group and package
